Move complex sync Action logic to OrderListSyncActionUseCase
